requirements.txt
google-genai
pandas
python-dotenv
matplotlib
streamlit
st-gsheets-connection